requirements.txt
dlib==19.20.0
idna==2.10
numpy==1.18.5
opencv-python==4.3.0.36
Pillow==7.2.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
torch==1.5.1
torchvision==0.6.1
